Add files via upload
This is a GUI version of the same tool. It does not share much code with the old version. Currently it does not have the ability to search for terms like the old one did, however it does let you step back and forth through files, and it let's you choose different directories without requiring a whole relaunch. See the readme in the root folder for a list of current problems.